Fix architecture tests - move EndpointResolver out of internal
