+++ core/sdk-core/src/main/java/software/amazon/awssdk/core/endpoint/EndpointResolver.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.core.internal.endpoint;
+package software.amazon.awssdk.core.endpoint;
 
 import software.amazon.awssdk.annotations.SdkProtectedApi;
 import software.amazon.awssdk.core.SdkRequest;
