+++ test/architecture-tests/archunit_store/18fc8858-1308-4d5d-b92d-87817d2fab53
@@ -31,7 +31,6 @@ Class <software.amazon.awssdk.imds.internal.AsyncHttpRequestHelper> depends on a
 Class <software.amazon.awssdk.imds.internal.AsyncHttpRequestHelper> depends on an internal API from a different module (Class <software.amazon.awssdk.core.internal.http.async.AsyncResponseHandler>)
 Class <software.amazon.awssdk.imds.internal.AsyncHttpRequestHelper> depends on an internal API from a different module (Class <software.amazon.awssdk.core.internal.http.async.SimpleHttpContentPublisher>)
 Class <software.amazon.awssdk.imds.internal.DefaultEc2MetadataAsyncClient> depends on an internal API from a different module (Class <software.amazon.awssdk.core.internal.http.loader.DefaultSdkAsyncHttpClientBuilder>)
-Class <software.amazon.awssdk.imds.internal.DefaultEc2MetadataClient> depends on an internal API from a different module (Class <software.amazon.awssdk.core.internal.http.loader.DefaultSdkHttpClientBuilder>)
 Class <software.amazon.awssdk.protocols.xml.AwsXmlProtocolFactory> depends on an internal API from a different module (Class <software.amazon.awssdk.core.internal.http.CombinedResponseHandler>)
 Class <software.amazon.awssdk.services.dynamodb.DynamoDbRetryPolicy> depends on an internal API from a different module (Class <software.amazon.awssdk.core.internal.retry.RetryPolicyAdapter>)
 Class <software.amazon.awssdk.services.dynamodb.DynamoDbRetryPolicy> depends on an internal API from a different module (Class <software.amazon.awssdk.core.internal.retry.SdkDefaultRetrySetting>)
@@ -49,8 +48,6 @@ Class <software.amazon.awssdk.transfer.s3.internal.GenericS3TransferManager> dep
 Class <software.amazon.awssdk.transfer.s3.internal.GenericS3TransferManager> depends on an internal API from a different module (Class <software.amazon.awssdk.services.s3.internal.multipart.MultipartDownloadResumeContext>)
 Class <software.amazon.awssdk.transfer.s3.internal.GenericS3TransferManager> depends on an internal API from a different module (Class <software.amazon.awssdk.services.s3.internal.multipart.MultipartDownloadUtils>)
 Class <software.amazon.awssdk.transfer.s3.internal.GenericS3TransferManager> depends on an internal API from a different module (Class <software.amazon.awssdk.services.s3.internal.multipart.MultipartS3AsyncClient>)
-Class <software.amazon.awssdk.transfer.s3.internal.GenericS3TransferManager> depends on an internal API from a different module (Class <software.amazon.awssdk.services.s3.internal.resource.S3AccessPointResource>)
-Class <software.amazon.awssdk.transfer.s3.internal.GenericS3TransferManager> depends on an internal API from a different module (Class <software.amazon.awssdk.services.s3.internal.resource.S3ArnConverter>)
 Class <software.amazon.awssdk.transfer.s3.internal.TransferManagerFactory> depends on an internal API from a different module (Class <software.amazon.awssdk.services.s3.internal.crt.S3CrtAsyncClient>)
 Class <software.amazon.awssdk.transfer.s3.internal.model.CrtFileUpload> depends on an internal API from a different module (Class <software.amazon.awssdk.services.s3.internal.crt.S3MetaRequestPauseObservable>)
 Class <software.amazon.awssdk.transfer.s3.internal.model.DefaultFileDownload> depends on an internal API from a different module (Class <software.amazon.awssdk.services.s3.internal.multipart.MultipartDownloadUtils>)
