+++ test/architecture-tests/archunit_store/4195d6e3-8849-4e5a-848d-04f810577cd3
@@ -11,9 +11,9 @@ Method <software.amazon.awssdk.core.retry.ClockSkew.getServerTime(software.amazo
 Method <software.amazon.awssdk.http.FileStoreTlsKeyManagersProvider.keyManagers()> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (FileStoreTlsKeyManagersProvider.java:52)
 Method <software.amazon.awssdk.http.SystemPropertyTlsKeyManagersProvider.keyManagers()> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (SystemPropertyTlsKeyManagersProvider.java:61)
 Method <software.amazon.awssdk.http.apache.ApacheHttpClient$ApacheConnectionManagerFactory.getSslContext(software.amazon.awssdk.utils.AttributeMap)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier)> in (ApacheHttpClient.java:699)
-Method <software.amazon.awssdk.http.apache5.Apache5HttpClient$ApacheConnectionManagerFactory.getSslContext(software.amazon.awssdk.utils.AttributeMap)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier)> in (Apache5HttpClient.java:741)
 Method <software.amazon.awssdk.http.apache.internal.RepeatableInputStreamRequestEntity.parseContentLength(java.lang.String)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier)> in (RepeatableInputStreamRequestEntity.java:113)
 Method <software.amazon.awssdk.http.apache.internal.utils.ApacheUtils.noSuchMethodThrownByNormalizeUriInvoker()> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier)> in (ApacheUtils.java:162)
+Method <software.amazon.awssdk.http.apache5.Apache5HttpClient$ApacheConnectionManagerFactory.getSslContext(software.amazon.awssdk.utils.AttributeMap)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier)> in (Apache5HttpClient.java:741)
 Method <software.amazon.awssdk.http.apache5.internal.RepeatableInputStreamRequestEntity.parseContentLength(java.lang.String)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier)> in (RepeatableInputStreamRequestEntity.java:131)
 Method <software.amazon.awssdk.http.apache5.internal.RepeatableInputStreamRequestEntity.parseContentType(java.lang.String)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier)> in (RepeatableInputStreamRequestEntity.java:143)
 Method <software.amazon.awssdk.http.nio.netty.internal.utils.NettyUtils.warnIfNotInEventLoop(io.netty.channel.EventLoop)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (NettyUtils.java:289)
@@ -36,16 +36,14 @@ Method <software.amazon.awssdk.protocols.xml.internal.unmarshall.XmlResponseHand
 Method <software.amazon.awssdk.regions.servicemetadata.EnhancedS3ServiceMetadata.profileFileSetting(java.util.function.Supplier, java.util.function.Supplier)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (EnhancedS3ServiceMetadata.java:118)
 Method <software.amazon.awssdk.services.s3.internal.endpoints.UseGlobalEndpointResolver.profileFileSetting(java.util.function.Supplier, java.util.function.Supplier)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (UseGlobalEndpointResolver.java:97)
 Method <software.amazon.awssdk.services.s3.internal.multipart.GenericMultipartHelper.cleanUpParts(java.lang.String, software.amazon.awssdk.services.s3.model.AbortMultipartUploadRequest$Builder)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (GenericMultipartHelper.java:121)
-Method <software.amazon.awssdk.services.s3.internal.multipart.KnownContentLengthAsyncRequestBodySubscriber.onSubscribe(org.reactivestreams.Subscription)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier)> in (KnownContentLengthAsyncRequestBodySubscriber.java:128)
-Method <software.amazon.awssdk.services.s3.internal.multipart.UploadWithUnknownContentLengthHelper$UnknownContentLengthAsyncRequestBodySubscriber.onSubscribe(org.reactivestreams.Subscription)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier)> in (UploadWithUnknownContentLengthHelper.java:143)
 Method <software.amazon.awssdk.services.s3.internal.s3express.UseS3ExpressAuthResolver.profileFileSetting(java.util.function.Supplier, java.util.function.Supplier)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (UseS3ExpressAuthResolver.java:80)
 Method <software.amazon.awssdk.services.s3.internal.settingproviders.DisableMultiRegionProviderChain.resolve()> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (DisableMultiRegionProviderChain.java:79)
 Method <software.amazon.awssdk.services.s3.internal.settingproviders.UseArnRegionProviderChain.resolveUseArnRegion()> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (UseArnRegionProviderChain.java:76)
 Method <software.amazon.awssdk.services.sqs.internal.batchmanager.ReceiveSqsMessageHelper.clear()> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (ReceiveSqsMessageHelper.java:132)
 Method <software.amazon.awssdk.transfer.s3.internal.AsyncBufferingSubscriber.onSubscribe(org.reactivestreams.Subscription)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier)> in (AsyncBufferingSubscriber.java:67)
 Method <software.amazon.awssdk.transfer.s3.internal.PauseResumeHelper.fileModified(software.amazon.awssdk.transfer.s3.model.ResumableFileUpload, software.amazon.awssdk.services.s3.S3AsyncClient)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (PauseResumeHelper.java:59)
-Method <software.amazon.awssdk.utils.ContentRangeParser.totalBytes(java.lang.String)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (ContentRangeParser.java:71)
 Method <software.amazon.awssdk.transfer.s3.progress.LoggingTransferListener.transferFailed(software.amazon.awssdk.transfer.s3.progress.TransferListener$Context$TransferFailed)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (LoggingTransferListener.java:76)
+Method <software.amazon.awssdk.utils.ContentRangeParser.totalBytes(java.lang.String)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (ContentRangeParser.java:71)
 Method <software.amazon.awssdk.utils.Logger.log(org.slf4j.event.Level, java.util.function.Supplier)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier)> in (Logger.java:205)
 Method <software.amazon.awssdk.utils.async.AddingTrailingDataSubscriber.onSubscribe(org.reactivestreams.Subscription)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (AddingTrailingDataSubscriber.java:73)
 Method <software.amazon.awssdk.utils.async.BaseSubscriberAdapter.onSubscribe(org.reactivestreams.Subscription)> calls method <software.amazon.awssdk.utils.Logger.warn(java.util.function.Supplier, java.lang.Throwable)> in (BaseSubscriberAdapter.java:99)
